Fix occasional double slashes in download URLs.
In some environments, the guessServerUrl() method could return an invalid URL containing two forward slashes "//" instead of one. This caused update downloads to fail in WP 4.6. See bug report:
https://github.com/YahnisElsts/plugin-update-checker/issues/66

The bug was in the *nix-like branch in guessServerUrl() that always added a forward slash to the path, even if there was one there already. Fixed by simplifying the method and only adding the slash only when necessary.