Add basic support for log rotation. Closes #46
This feature is off by default. To enable it, call enableLogRotation() and specify the rotation period and the number of old log files to keep. Example: 

$server->enableLogRotation(Wpup_UpdateServer::FILE_PER_MONTH, 6);

This is a very basic, limited implementation of log rotation. I would recommend using logrotate instead. It has more options and is likely to perform better.