Documentation: Give the server all query params
The example code in the "Running the server from another script" should pass all available query parameters to the `handleRequest` method, not just `action` and `slug`. The update server will *work* with just those two parameters, but you need to pass in a bunch of other stuff if you want it to log things like the installed version, PHP version and so on.